# Use FMA instruction in CpuMath for .NET Core 3