… DotnetCLIVersion.txt