Fix https://github.com/dotnet/machinelearning/issues/832

Test with `..\..\Tools\dotnetcli\dotnet.exe run -c Release-Intrinsics --allCategories=Fma`

Before:

```
// * Summary *

BenchmarkDotNet=v0.11.1, OS=Windows 10.0.17758
Intel Core i9-7980XE CPU 2.60GHz (Max: 2.59GHz), 1 CPU, 36 logical and 18 physical cores
.NET Core SDK=3.0.100-alpha1-009642
  [Host] : .NET Core 3.0.0-preview1-27004-04 (CoreCLR 4.6.27003.04, CoreFX 4.6.27003.02), 64bit RyuJIT

Toolchain=InProcessToolchain  

         Method |       Mean |      Error |     StdDev |     Median |
--------------- |-----------:|-----------:|-----------:|-----------:|
      AddScaleU |   271.8 us |  0.3131 us |  0.2614 us |   271.9 us |
     AddScaleSU | 2,890.7 us | 35.0849 us | 32.8185 us | 2,896.3 us |
  AddScaleCopyU |   396.8 us |  2.2882 us |  2.1404 us |   395.7 us |
         SumSqU |   195.7 us |  0.3503 us |  0.3277 us |   195.6 us |
     SumSqDiffU |   194.3 us |  0.3320 us |  0.2592 us |   194.2 us |
           DotU |   268.2 us |  7.1498 us | 20.7429 us |   274.9 us |
          DotSU | 2,355.5 us | 47.0184 us | 46.1783 us | 2,357.3 us |
          Dist2 |   276.2 us |  0.5637 us |  0.5272 us |   276.3 us |
  SdcaL1UpdateU |   481.4 us |  4.9104 us |  4.5932 us |   480.9 us |
 SdcaL1UpdateSU | 8,043.5 us | 44.7203 us | 39.6434 us | 8,030.6 us |
        MatMulX |   137.2 us |  2.9038 us |  8.1903 us |   139.6 us |
```
After:

```
// * Summary *

BenchmarkDotNet=v0.11.1, OS=Windows 10.0.17758
Intel Core i9-7980XE CPU 2.60GHz (Max: 2.59GHz), 1 CPU, 36 logical and 18 physical cores
.NET Core SDK=3.0.100-alpha1-009642
  [Host] : .NET Core 3.0.0-preview1-27004-04 (CoreCLR 4.6.27003.04, CoreFX 4.6.27003.02), 64bit RyuJIT

Toolchain=InProcessToolchain  

         Method |        Mean |      Error |     StdDev |      Median |
--------------- |------------:|-----------:|-----------:|------------:|
      AddScaleU |   273.38 us |  0.4326 us |  0.4046 us |   273.47 us |
     AddScaleSU | 2,917.50 us | 33.6275 us | 31.4552 us | 2,911.38 us |
  AddScaleCopyU |   393.58 us |  1.0500 us |  0.9822 us |   393.61 us |
         SumSqU |   195.53 us |  1.6073 us |  1.5035 us |   194.86 us |
     SumSqDiffU |   196.62 us |  0.1949 us |  0.1824 us |   196.59 us |
           DotU |   198.04 us |  0.3970 us |  0.3713 us |   198.12 us |
          DotSU | 2,337.79 us | 12.5178 us | 11.7092 us | 2,337.42 us |
          Dist2 |   264.65 us |  7.0762 us | 19.7257 us |   271.17 us |
  SdcaL1UpdateU |   481.08 us |  1.3783 us |  1.2218 us |   480.94 us |
 SdcaL1UpdateSU | 8,001.56 us | 28.4807 us | 26.6409 us | 7,990.17 us |
        MatMulX |    96.01 us |  0.4207 us |  0.3729 us |    95.99 us |

```

@eerhardt @tannergooding PTAL