@tannergooding 2018-10-18
> Intel Core i9-7980XE CPU 2.60GHz (Max: 2.59GHz), 1 CPU, 36 logical and 18 physical cores

It would be nice to test this on some hardware that normal users would be more likely to have (I believe we generally test on a 4-core i7).