@tannergooding 2018-10-18
> Which generation(s) of platforms (e.g. Coffee Lake, Sky Lake etc.) do you use for testing?

My current work box is an `i7 Kaby Lake` (`i7 7700 @ 3.60GHz`). I am still attempting to determine the exact micro-architecture (or range of micro-architectures) we are using for our official benchmarks. -- CC. @adamsitnik who might have some idea.