@adamsitnik 2018-10-18
@jorive what is our hardware setup for the perf lab? and what machines did we order some time ago?