@jorive 2018-10-18
@adamsitnik We have been running .NET Core benchmarks on Haswell machines (i7-4790).
Now, with respect to the hardware we ordered a while back I do not know what we are getting. @brianrob and @maririos might have more information.