@maririos 2018-10-19
> what machines did we order some time ago?

We are in the process of revisiting the orders and the current need. Ping me if you want to be involved in the conversation for that new hardware.