@adamsitnik 2018-10-19
@tannergooding do you know if there are any env vars to make `Fma.IsSupported` return false for the purpose of testing?