@tannergooding 2018-10-19
> @tannergooding do you know if there are any env vars to make Fma.IsSupported return false for the purpose of testing?

We have one, but it is currently only enabled for 'checked' builds of the Runtime and is not available as a "retail" flag. It has been brought up before that we may need it as a retail flag for cases like this.
CC. @CarolEidt, @danmosemsft, @eerhardt, @fiigii