@fiigii 2018-10-19
Opened an issue at https://github.com/dotnet/coreclr/issues/20498, I will expose these knobs in release build once it gets approved.